requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic-settings
strawberry-graphql
httpx
openai
pytest
